requirements.txt
flask==3.0.0
flask-cors
pandas
fpdf
google-generativeai
python-dotenv